consumer/requirements.txt
pika
retrying
requests
mysql-connector-python